Enable ADB over network on cic
Basing on the requirement of epic of OAM-87737, we need to enable
secure ADB on cic.
Since we usually connect to cic devices by ADB over network, adb
daemon needs to listen on a network port (Here it's 5555). So
the adb authentication pop-up dialog can be displayed by doing this:

	adb connect localhost:5555

Tracked-On: OAM-89667
Signed-off-by: ji, zhenlong z <zhenlong.z.ji@intel.com>